Make IntervalConfigNotifier publicly available
Moves the IntervalConfigNotifier and its provider to be usable by other
parts of the app, which allows other providers and widgets to use its
state. This simplifies the code as the interval will not need to be
passed around explicitly among widgets and providers. Most importantly,
this allows other providers to fetch the configured interval on their
own, i.e. not be 'family' providers any more.