Allow going back from interval config page
Shows an app bar in that page, now with a title, and will show the
'back' button when a previous page exists.

Also makes IntervalConfig provider an auto-disposed one. This will be
necessary when more modes are supported and e.g. navigating back from
interval config page - when going to interval config again, the UI
should be reset.