Unify duration formatting
There is now only one main parameterized Duration formatter. Depending
on the needs, the UI layer will call this formatter with different
parameters.

Now milliseconds are a separate component in the UnpackedDuration
structure as it gets dedicated handling. This makes it easier to print
only what's really needed, and does not rely on double formatting.

This change also adds the first unit tests in this project, rejoice.