Modularize workout page
TimerPage has been renamed to WorkoutPage, and it is now just a
skeleton with placeholders for the top and bottom widgets, which are
specified as parameters whenever the WorkoutPage is created.

This change will make it easier to create new modes, which will not have
rounds ('interval' mode), or display timing information in a different
way, like 'AMRAP' (countdown) vs 'stopwatch' (counting up).