Implement a fancier workout screen
Now when the timer is running, the top and bottom widgets take up half
of the screen each, and workout menu is invisible. When the timer is not
running, the menu appears at the very bottom, and main widgets are made
a bit smaller and stay anchored to the top of the screen.

The opacity of the 'round summary' button is now animated as well.

Finally, the color of the text displayed by the 'timer widget' (elapsed
or remaining time, depending on the mode) is also changed with an
animation.