Add support for input of multiple intervals
AMRAP uses the 'ongoing' definition, which plays nicely with it being
preserved when going back. 'interval' model will use this new
functionality in near future. For now, TimerNotifier cannot deal with
more than 1 either, so the first interval (which is always the 'ongoing'
one due to how the code is written) is always used (or none, in the
'stopwatch' mode).