Simplify OngoingIntervalDefinition API slightly
As an 'empty' definition should never actually be created (is null in
the state), this API had the potential to be confusing. In reality, the
sole purpose of the 'isEmpty' property was to be overridden in the input
page 'null' object. As it is only used there, let's define it there and
not pollute the API surface. It is now defined as a private extension -
actually, 'isNotEmpty' is defined as it is always used this way.