Split interval input and intervals setup
A new provider is added which keeps track of all added intervals, it is
used by timer provider to read them for the workout. Interval input adds
its interval to this new provider. In the future, it will be possible to
remove, update or reorder intervals in the upcoming 'intervals' mode.