Refactor IntervalDefinition and interval input
There is now less code and it is more flexible at the cost of allowing
creation of a '0 time' interval (all time components set to 0), which is
not actually an issue.

It introduces the idea of a 'prototype' IntervalDefinition which will
be used to create new instances when input changes. This will make
everything easier once more fields are introduced, e.g. repetitions -
they should be preserved when input changes during interval update, but
it wouldn't be the case with the current code and special handling would
need to be introduced.

This also allows simplification of the special 'null-object' deffinition
in input page.