Split IntervalsSetup model from IntervalGroup
Now setup has its own model that is converted to IntervalGroup. This
makes more assertions in IntervalGroup possible and is cleaner in
general. This is also the first step for support of multiple interval
groups.

This commit leaves the code in a bit of a filthy state which is
something in between not supporting groups and supporting them.