Update whatever we can
AGP cannot be updated due to 'wakelock' and 'audioplayers' not using the
mandatory 'namespace' property, we need to wait.

Kotlin cannot be updated due to an issue with inconsistent Java versions
being set for various tasks, which will be fixed in AGP 8.1. But AGP
cannot be updated yet, see above.

riverpod is not updated yet as it will not compile due to many breaking
changes, and it is not necessary to update, the projects still builds
and works fine. This will be done in a separate change.